The work minus the data
So this includes the entire function minus a true dataset, I have a sample A slotted in right now. All that is required is putting in our own dataset and naming it A BEFORE we hit this display(which is at the end of the file where it gets all the inputs from the user as desired).